Docker version update, some fixes on validation on the docker entrypoints of amavis, docs and some other tweaks

+++ amavis/docker-entrypoint.sh
@@ -4,6 +4,28 @@ set -m -o pipefail
 # This script is part of MailD
 # Copyright 2020-2024 Pavel Milanes Costa <pavelmc@gmail.com>
 
+# preconditions on start: IP of some of the hosts
+CLAMAVIP=`host ${CLAMAV} | awk '/has address/ { print $4 }'`
+MTAIP=`host ${MTA} | awk '/has address/ { print $4 }'`
+CRONIP=`host ${CRON} | awk '/has address/ { print $4 }'`
+
+# check if  any of the IPs are empty
+if [ -z "${CLAMAVIP}" ] ; then
+    echo "====== !!!!!!!!!!!!!!!!!! ======="
+    echo "CLAMAV IP is empty: die"
+    exit 1
+fi
+if [ -z "${MTAIP}" ] ; then
+    echo "====== !!!!!!!!!!!!!!!!!! ======="
+    echo "MTA IP is empty: die"
+    exit 1
+fi
+if [ -z "${CRONIP}" ] ; then
+    echo "====== !!!!!!!!!!!!!!!!!! ======="
+    echo "CRON IP is empty: die"
+    exit 1
+fi
+
 # copy or overwrite the config files from the default ones
 cd /etc/amavis
 rm -rdf conf.d
@@ -16,29 +38,9 @@ echo "POSTGRES_DB=${POSTGRES_DB}" >> "${CFILE}"
 echo "POSTGRES_USER=${POSTGRES_USER}" >> "${CFILE}"
 echo "POSTGRES_PASSWORD=${POSTGRES_PASSWORD}" >> "${CFILE}"
 echo "MTA=${MTA}" >> "${CFILE}"
-MTAIP=`host ${MTA} | awk '/has address/ { print $4 }'`
 echo "MTAIP=${MTAIP}" >> "${CFILE}"
-CLAMAVIP=`host ${CLAMAV} | awk '/has address/ { print $4 }'`
 echo "CLAMAVIP=${CLAMAVIP}" >> "${CFILE}"
-CRONIP=`host ${CRON} | awk '/has address/ { print $4 }'`
 echo "CRONIP=${CRONIP}" >> "${CFILE}"
-
-# check if  any of the IPs are empty
-if [ -z "${CLAMAVIP}" ] ; then
-    echo "====== !!!!!!!!!!!!!!!!!! ======="
-    echo "CLAMAV IP is empty"
-    exit 1
-fi
-if [ -z "${MTAIP}" ] ; then
-    echo "====== !!!!!!!!!!!!!!!!!! ======="
-    echo "MTA IP is empty"
-    exit 1
-fi
-if [ -z "${CRONIP}" ] ; then
-    echo "====== !!!!!!!!!!!!!!!!!! ======="
-    echo "CRON IP is empty"
-    exit 1
-fi
 
 # IP data for the checks
 echo $CLAMAVIP > /tmp/CLAMAVIP
@@ -107,7 +109,7 @@ function get_domains() {
     # query to get the domains
     QUERY="SELECT domain FROM domain;"
 
-    # craaft the auth credentials & secure it
+    # craft the auth credentials & secure it
     echo "$POSTGRES_HOST:5432:$POSTGRES_DB:$POSTGRES_USER:$POSTGRES_PASSWORD" > ~/.pgpass
     chmod 0600 ~/.pgpass &1>2
 
@@ -117,7 +119,7 @@ function get_domains() {
     # validate
     R=$?
     if [ ! $R -eq 0 ] ; then
-        echo "Error, could not connect to database"
+        echo "Error, could not connect to database" >&2
         exit 1
     fi
 
@@ -131,8 +133,8 @@ function get_domains() {
     #  domain  
     #----------
     # ALL
-    # sample1.com.jm
-    # exercises.jm
+    # sample1.com.tld
+    # exercises.tld
     #(2 rows)
 
     # match any domain like string on the results
@@ -157,11 +159,9 @@ if [ "${DKIM_SIGNING}" ] ; then
     DKIM_DOMAINS=$(get_domains)
     FILESIGN=/etc/amavis/conf.d/22-dkim_signing
 
-    # debug dkim_domians if debugging
-    if [ "${AMAVIS_DEBUG}" ] ; then
-        echo "DKIM_DOMAINS: ${DKIM_DOMAINS}"
-        echo "FILESIGN: ${FILESIGN}"
-    fi
+    # debug dkim_domians
+    echo "DKIM_DOMAINS: ${DKIM_DOMAINS}"
+    echo "FILESIGN: ${FILESIGN}"
 
     # setup only if there are domains to process
     if [[ "${DKIM_DOMAINS}" ]] ; then
